requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
absl-py==1.2.0; python_version >= '3.6'
appnope==0.1.3; sys_platform == 'darwin'
asttokens==2.0.8
backcall==0.2.0
cachetools==5.2.0; python_version ~= '3.7'
certifi==2022.9.14; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
cycler==0.11.0; python_version >= '3.6'
decorator==5.1.1; python_version >= '3.5'
executing==1.0.0
fonttools==4.37.2; python_version >= '3.7'
google-auth-oauthlib==0.4.6; python_version >= '3.6'
google-auth==2.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
grpcio==1.49.0; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
importlib-metadata==4.12.0; python_version < '3.10'
ipython==8.5.0
jedi==0.18.1; python_version >= '3.6'
kiwisolver==1.4.4; python_version >= '3.7'
markdown==3.4.1; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
matplotlib-inline==0.1.6; python_version >= '3.5'
matplotlib==3.5.3
numpy==1.23.3
oauthlib==3.2.1; python_version >= '3.6'
opencv-python==4.6.0.66
packaging==21.3; python_version >= '3.6'
pandas==1.4.4
parso==0.8.3; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==9.2.0
prompt-toolkit==3.0.31; python_full_version >= '3.6.2'
protobuf==3.19.5; python_version >= '3.5'
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1-modules==0.2.8
pyasn1==0.4.8
pygments==2.13.0; python_version >= '3.6'
pyparsing==3.0.9; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2022.2.1
pyyaml==6.0
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.28.1
rsa==4.9; python_version >= '3.6'
scipy==1.9.1
seaborn==0.12.0
setuptools==65.3.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
stack-data==0.5.0
tensorboard-data-server==0.6.1; python_version >= '3.6'
tensorboard-plugin-wit==1.8.1
tensorboard==2.10.0
thop==0.1.1.post2209072238
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
traitlets==5.4.0; python_version >= '3.7'
typing-extensions==4.3.0; python_version >= '3.7'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
wcwidth==0.2.5
werkzeug==2.2.2; python_version >= '3.7'
wheel==0.37.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
zipp==3.8.1; python_version >= '3.7'
